Extract stat details to a component
